massage Travis configuration some more

+++ deploy.sh
@@ -8,9 +8,12 @@ mv src/index.html out
 sysconfcpus/bin/sysconfcpus -n $CPUNUM elm-make src/Main.elm --output out/elm.js --yes
 cd out
 
-git init
-git config user.name "Travis CI"
-git config user.email "jvoigtlaender@users.noreply.github.com"
-git add .
-git commit -m "Travis deploy to gh-pages"
-git push --force --quiet "https://$GH_TOKEN@github.com/$TRAVIS_REPO_SLUG" master:gh-pages >/dev/null 2>&1
+if [[ "$TRAVIS_PULL_REQUEST" == "false" && "$TRAVIS_BRANCH" == "master" ]];
+then
+  git init --quiet;
+  git config user.name "Travis CI";
+  git config user.email "jvoigtlaender@users.noreply.github.com";
+  git add .;
+  git commit -m "Travis deploy to gh-pages";
+  git push --force --quiet "https://$GH_TOKEN@github.com/$TRAVIS_REPO_SLUG" master:gh-pages >/dev/null 2>&1;
+fi
